+++ CHANGELOG.md
@@ -2,6 +2,8 @@ TBD
 ===
 Unreleased changes. Release notes have not yet been written.
 
+* add extern crate memchr into memchr bench
+
 0.1.1 (2021-05-14)
 =====